refactor: consolidate MCP tool-name parsing into internal/mcpname
The mcp__<server>__<function> / MCP: parsing introduced in AGE-2749 lived
in the risk_analysis Temporal-activity package, but the same SplitN("__")
logic was hand-rolled in 5 more places in the hooks package (shadow-mcp
guard + per-host telemetry attribution), with subtle drift in empty-part
handling.

Extract IsMCPToolName / MCPFunctionOf / MCPServerOf / AttributeTool into a
new dependency-free internal/mcpname package and route every caller through
it: risk_analysis (llm_judge, analyze_batch) and hooks (mcp_match,
shadow_mcp_access, codex_hooks, cursor_hooks, pending_helpers).

cursor_hooks keeps an explicit mcp__ guard so the shared parser's Cursor
"MCP:" branch does not collide with cursor's dedicated MCP: handling.
